SPS: Move creation of scaling tree to a new service
