+++ com.vrcfury.vrcfury/Editor/VF/Service/ScalePropertyCompensationService.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f860e5e13a768044ab74f0d6d4c31855
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
